Convert JS to TS
