+++ client/src/index.js
@@ -6,8 +6,8 @@ import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
-  </React.StrictMode>,
+  <App />
+  < /React.StrictMode>,
   document.getElementById('root')
 );
